add address type and types.go
